Big-data/de2_heart_disease/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
mlxtend>=0.23.0
statsmodels>=0.14.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
pyarrow>=13.0.0
pyyaml>=6.0
joblib>=1.3.0
tqdm>=4.65.0
shap>=0.44.0
papermill>=2.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
kaggle>=1.5.0
